Decided to make the windows form resizable. The small size was nice and compact, but now there's room to display a grid of matched DMC values.
Added a grid to display each DMC value. The grid represents the pixels of the image

Fixed aspect ratio not being correct. Had to cast the height and width to floats. As dividing an int by a larger int gave 0. Which would result in divisions by 0 when calculating the resized height of the image.